Added a few small items such as reformatting my food table and adding some hyperlinks.
I added a table to the food list so that it looks nicer and easier on the eyes.  I also made the question mark on the welcome page a hyperlink so I can go to the food table.  Also, when the user uses the random button and gets their random restaurant, I added a back button so they can go back to the welcome page.